% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create.FRBS.R
\name{Create.FRBS.R}
\alias{Create.FRBS.R}
\title{Function to create a zero-order Takagi–Sugeno–Kang fuzzy rule-based systems (FRBS)}
\usage{
Create.FRBS.R(ImpVariables, Range, MFfunction, MFparameters, Consequents)
}
\arguments{
\item{ImpVariables}{Rhe names of the input variables in vector format}

\item{Range}{a matrix with as many columns as input variables with the minimum and maximum expected values for each variable}

\item{MFfunction}{the selected membership function}

\item{MFparameters}{the parameters of the membership functions. It should be a matrix with as many columns as fuzzy sets. The columns for each variable must include the variable name as indicated in ImpVariables.}

\item{Consequents}{the fuzzy rules consequents.}
}
\value{
an R list containing the five components of the FRBSs
}
\description{
Function to create a zero-order Takagi–Sugeno–Kang fuzzy rule-based systems (FRBS)
}
