% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembershipFunctions.R
\name{SIGMF}
\alias{SIGMF}
\title{Sigmoidal membership functions}
\usage{
SIGMF(pattern, parameters)
}
\arguments{
\item{pattern}{the values of the input variable for whom the membership is calculated}

\item{parameters}{the parameters of the membership function}
}
\value{
membership of the evaluated values
}
\description{
Sigmoidal membership functions
}
